refactor(convert): extract background from extract_block_style
